Logger Moved into Core
