+++ Core/CtrlMiscUtils/CtrlMiscUtils.vcxproj.filters
@@ -15,10 +15,10 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Include\ThreadSafeVector.h">
+    <ClInclude Include="ThreadSafeQueue.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Include\ThreadSafeQueue.h">
+    <ClInclude Include="ThreadSafeVector.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
