+++ Core/GreatBath/FilePathProcessing.cpp
