+++ Core/GreatBath/Logger.cpp
