+++ Ctrl 2K Engine 0.0.1.sln
@@ -25,8 +25,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Material", "Ctrl 2K Engine 
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial", "Ctrl 2K Engine 0.0.1\SunDial.vcxproj", "{EB11A967-DAB8-4DE2-894C-F5014E62137E}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GreatBath", "Ctrl 2K Engine 0.0.1\GreatBath.vcxproj", "{55BB3977-E4AA-4C08-B457-B1214977396C}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SunDial", "SunDial", "{A1C2A35E-4285-470E-AD90-009320CE8212}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SunDial Profiler", "Ctrl 2K Engine 0.0.1\SunDial Profiler.vcxproj", "{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}"
@@ -35,11 +33,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assets", "Assets", "{26927F
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AssetsCore", "Ctrl 2K Engine 0.0.1\AssetsCore.vcxproj", "{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CtrlCore", "CtrlCore", "{EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Petra", "Core\Petra\Petra.vcxproj", "{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CtrlMiscUtils", "Core\CtrlMiscUtils\CtrlMiscUtils.vcxproj", "{385BC729-58FA-49D3-8609-0A33B280E111}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GreatBath", "Core\GreatBath\GreatBath.vcxproj", "{55BB3977-E4AA-4C08-B457-B1214977396C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -97,14 +97,6 @@ Global
 		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x64.Build.0 = Release|x64
 		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.ActiveCfg = Release|Win32
 		{EB11A967-DAB8-4DE2-894C-F5014E62137E}.Release|x86.Build.0 = Release|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.ActiveCfg = Debug|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.Build.0 = Debug|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.ActiveCfg = Debug|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.Build.0 = Debug|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.ActiveCfg = Release|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.Build.0 = Release|x64
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.ActiveCfg = Release|Win32
-		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.Build.0 = Release|Win32
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.ActiveCfg = Debug|x64
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x64.Build.0 = Debug|x64
 		{75CDE69C-F1E7-4CE0-B20D-65E9C5A9DBE1}.Debug|x86.ActiveCfg = Debug|Win32
@@ -137,6 +129,14 @@ Global
 		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x64.Build.0 = Release|x64
 		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.ActiveCfg = Release|Win32
 		{385BC729-58FA-49D3-8609-0A33B280E111}.Release|x86.Build.0 = Release|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.ActiveCfg = Debug|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x64.Build.0 = Debug|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.ActiveCfg = Debug|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Debug|x86.Build.0 = Debug|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.ActiveCfg = Release|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x64.Build.0 = Release|x64
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.ActiveCfg = Release|Win32
+		{55BB3977-E4AA-4C08-B457-B1214977396C}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -148,6 +148,7 @@ Global
 		{FC37A3E4-4FBE-4088-9DE8-2318C80E81E6} = {26927F71-5167-41ED-8F46-A20AA59F0B40}
 		{15AD7AFD-0DDB-4B7C-823F-3B54CF4D41AF} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 		{385BC729-58FA-49D3-8609-0A33B280E111} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
+		{55BB3977-E4AA-4C08-B457-B1214977396C} = {EFEC0CF7-F3B7-4C5B-A685-2441C61A6147}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {70BAD174-8BA7-4F92-9815-19B71A819840}
