+++ Ctrl 2K Engine 0.0.1/Material.vcxproj
@@ -163,9 +163,6 @@
     <ProjectReference Include="AssetsCore.vcxproj">
       <Project>{fc37a3e4-4fbe-4088-9de8-2318c80e81e6}</Project>
     </ProjectReference>
-    <ProjectReference Include="GreatBath.vcxproj">
-      <Project>{55bb3977-e4aa-4c08-b457-b1214977396c}</Project>
-    </ProjectReference>
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
