+++ Ctrl 2K Engine 0.0.1/Source/Material/TexturesInternal.cpp
@@ -11,7 +11,7 @@
 
 #include "SOIL.h"
 
-#include "Logger.h"
+#include "../../../Core/GreatBath/Logger.h"
 #include "TexturesInternal.h"
 #include "AssetsManager.h"
